#wizard_ver3: Implementacion del wizard usando router
