+++ src/app/(pages)/drugs-wizard/step1/page.tsx
@@ -1,22 +1,33 @@
+'use client'
+
 import { useForm, FormProvider } from "react-hook-form"
 import { useContext } from "react";
-import { WizardContext } from "./wizard-context";
+import { WizardContext } from "../wizard-context";
 import MainForm from "@/app/components/main-form";
 import { Box, Button } from "@mui/material";
-import asas from "../../constants/asas.json";
+import asas from "../../../constants/asas.json";
+import { useRouter } from "next/navigation";
 import NavigateNextIcon from '@mui/icons-material/NavigateNext';
 
 export default function DrugWizardStep1() {
     const mainForm = useForm({ mode: "onBlur" });
     const { watch, handleSubmit } = mainForm;
 
-    const { asaObj, weightObj, stepObj } = useContext(WizardContext);
+    const { asaObj, weightObj,drugObj } = useContext(WizardContext);
     const [, setWeightState] = weightObj;
     const [selectedAsa, setSelectedAsa] = asaObj;
-    const [, setActiveStep] = stepObj;
+    const [drugs] = drugObj;
+
+    const router = useRouter()
 
     const { weight } = watch();
+
+    // if(drugs.length > 0){
+    //     alert("drugs tiener dro")
+    // }
     setWeightState(weight)
+    
+    
     const { asa } = watch();
     const filteredSelectedAsa = asas.find(a => a.id === asa);
     setSelectedAsa(filteredSelectedAsa);
@@ -37,7 +48,7 @@ export default function DrugWizardStep1() {
             >
                 <Button
                     startIcon={<NavigateNextIcon />}
-                    variant="contained" onClick={handleSubmit(() => setActiveStep(1))} >
+                    variant="contained" onClick={handleSubmit(() => router.push('./step2'))} >
                     Siguiente
                 </Button>
             </Box>
